refactor multiwindow tests
